Add a convertor from old timeseries to new ones
This is necessary to run this on an existing old database in
order to be able to use the new Timeserieswidget.

The script runs on all asset attributes which can be provided in
timeseries format and update the format of the dumped json in
the text field in the database.

The variable `save` has to manually be set to True once you
decide to affect the database with the new format